Color themes are here!
Added a colors.cfg file that has a light and dark theme. Either can be
set to the default. I also tuned how the stage box colors are done, so
it's more like Jenkins (i.e. if the job is unstable, the job box and the
stage that was unstable will display in that color, successful stages
will display in the success color. Yellow and Green respectively by
default)

Updated the README as well.